Update patch set 1
Patch Set 1: Code-Review-1

We need to decide whether or not to use sony_common. As of now, the only devices that depend on this repo (in 5.0/1) are the fusion3s. It would be pointless to keep it there in that case (might as well move it to msm8960 or even fusion3-common).
Should you decide to go for sony_common, you should remove some dupes (e.g. extract_elf_ramdisk)

BTW: What about the other Sony devices we supported in 4.4, say rhine for example?

Patch-set: 1
Label: Code-Review=-1