Update patch set 1
Patch Set 1:

> there is inconsistency with TWRP and other roms.
Which inconsistency? I have just visited some repos (google, samsung, ...) and they either define ext4 or both.
 > Mostly if you wipe from rom, partition will be ext4 formated and
 > some more compatibility problem popup into twrp (not able to mount
 > storage, etc...)
Could you send me detailed instruction to reproduce this? The implication of this change is that one MUST explicitly format partitions with f2fs, otherwise one can't boot. The fstab as it is now is intended such that the user is able to use partitions formatted with either fs.

 > Also f2fs is better than ext4.
This is a highly subjective thing. f2fs is designed for flash storage, but ext4 is longer around and mature. It also has received several enhancements w.r.t. flash storage.

Patch-set: 1