Update patch set 4
Patch Set 4:

> are you sure this is not the issue with always a denial at the
 > first query?

Yes. As discussed, the "denial at the first query" resulted from duplicate sepolicy rules that are not by the policy "compiler". Hence, it's a policy thing: either we have some duplicate somewhere, or this one is just not included (whereby the latter is more likely).

Correcting the duplicate detection would be good, but is probably out-of-scope for this project.

Patch-set: 4
Label: Code-Review=0