Update patch set 1
Patch Set 1: Verified+1

> Here it is : https://github.com/LineageOS/android_device_sony_huashan/commit/cc7131f40d7e98e5fa4031584e271342c556acb6
 > with links to https://source.android.com/devices/storage/adoptable.html
 > & https://source.android.com/devices/storage/config.html#android_6_0

It's probably more a cosmetic change, since the documentation only discusses the vold behavior. I'd assume that only the first chars are parsed such that the behavior is equivalent for "usb" and "usbdisk"...

Anyways, I'd merge this when the author is corrected (David) because we are in line with the docs then.

Patch-set: 1
Reviewer: Gerrit User 1001607 <1001607@56841c99-7c94-43d4-bf0b-cb4707e66d55>
Label: Verified=+1